LayoutTests/http/tests/media/track-in-band-hls-metadata-expected.txt

  
Test for metadata tracks from Apple HLS stream.


** Set video.src, wait for media data to load

** Set video.src, wait for media data to load
RUN(video.src = 'http://127.0.0.1:8000/media/resources/hls/metadata/prog_index.m3u8')

EVENT(addtrack)
EXPECTED (tracks.length == '1') OK
RUN(track = video.textTracks[0])
EXPECTED (track.kind == 'metadata') OK
EXPECTED (track.mode == 'disabled') OK
RUN(track.mode = 'hidden')
END OF TEST
